@@ -12,37 +12,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "6.0.9",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
+    "@angular/core": "6.0.9",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/router": "6.0.9",
     "aos": "^2.2.0",
-    "ap-angular2-fullcalendar": "^1.3.7",
+    "ap-angular-fullcalendar": "1.6.4",
     "bootstrap": "4.1.2",
-    "contentful": "^5.1.3",
+    "contentful": "6.1.1",
     "core-js": "^2.4.1",
-    "fullcalendar": "3.6.1",
     "jquery": "3.0.0",
     "marked": "^0.3.19",
     "moment": "^2.22.0",
-    "ng-fullcalendar": "^1.5.5",
     "ngx-md": "^3.1.1",
     "popper.js": "^1.14.1",
     "request": "^2.87.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.2.2",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.9",
+    "@angular/language-service": "6.0.9",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
+    "@types/jquery": "2.0.47",
     "@types/marked": "^0.3.0",
     "@types/node": "~6.0.60",
     "codelyzer": "^4.0.1",
@@ -56,6 +57,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "2.7.2"
   }
 }
